chore(route): rename route file
- rename routes file to index.js

[Finishes #159381385]
