+++ UI/main/profile.html
@@ -116,5 +116,6 @@
 	</footer>
 
 	<script  src="../scripts/nav.js"></script>
+	<script  src="../scripts/profile.js"></script>
 </body>
 </html>
